python-function/fastAPI/requirements.txt
fastapi
uvicorn

opencv-contrib-python==4.6.0.66
requests
numpy
Pillow
python-multipart